@Home, Added packages for 100 coins($15), 250 coins($30).

+++ app/src/main/java/apimodels/User.java
@@ -88,7 +88,7 @@ public class User
         @SerializedName("id")
         private int id;
         @SerializedName("rating")
-        private int rating;
+        private float rating;
         @SerializedName("last_login")
         private String lastLogin;
         @SerializedName("email")
@@ -205,12 +205,12 @@ public class User
             this.id = id;
         }
 
-        public int getRating()
+        public float getRating()
         {
             return rating;
         }
 
-        public void setRating(int rating)
+        public void setRating(float rating)
         {
             this.rating = rating;
         }
